draft fix for issue #6775
timeSequence() was sometimes producing wrong times in DST time zones. 
This was due to 'wrong' (or rather, inconsistent with others) handling of 
invalid times.

This fix modifies timeDate() to produce shift such times by +1 hour to make 
them  valid. This is sensible for timeSequence() but not necessarilly in 
general. A final fix may require addition of an argument to timeDate() and/or
the internal function .formatFinCenterNum() handling that. Also, the default may have to be allowing the non-existent date-times, as in strptime().




git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rmetrics@6335 5d4a7380-5c0c-0410-b446-d4a5018c0119